+++ lib/helpers.js
@@ -1,12 +1,10 @@
 'use strict';
 
 const logger = require('./logger');
-const { PhoneNumberFormat, PhoneNumberUtil } = require('google-libphonenumber');
 
 const config = require('../config/lib/helpers');
 const helpers = require('./helpers/index');
 const InternalServerError = require('../app/exceptions/InternalServerError');
-const UnprocessibleEntityError = require('../app/exceptions/UnprocessibleEntityError');
 
 // TODO: Move contents of this helper.js file into lib/helpers/index.js or to modular helpers
 
@@ -88,17 +86,4 @@ module.exports.sendResponseWithMessage = function (res, message) {
 module.exports.addBlinkSuppressHeaders = function addBlinkSuppressHeaders(res) {
   logger.debug('Adding Blink suppress headers', {}, res);
   return res.setHeader('x-blink-retry-suppress', true);
-};
-
-module.exports.formatMobileNumber = function formatMobileNumber(mobile, format = 'E164', countryCode = 'US') {
-  logger.debug('formatMobileNumber params', { mobile });
-  const phoneUtil = PhoneNumberUtil.getInstance();
-  const phoneNumberObject = phoneUtil.parse(mobile, countryCode);
-  if (!phoneUtil.isValidNumber(phoneNumberObject)) {
-    const error = new UnprocessibleEntityError('Cannot format mobile number.');
-    throw error;
-  }
-  const result = phoneUtil.format(phoneNumberObject, PhoneNumberFormat[format]);
-  logger.debug('formatMobileNumber return', { result });
-  return result;
 };
