+++ lib/helpers/front.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const logger = require('../logger');
 const frontClient = require('../front');
-const requestHelper = require('./request');
+const helpers = require('../helpers');
+const logger = require('../logger');
 
 module.exports = {
   getConversationByUrl: function getConversationByUrl(url) {
@@ -18,14 +18,24 @@ module.exports = {
    */
   parseBody: function parseBody(req) {
     const body = req.body;
-    requestHelper.setOutboundMessageText(req, body.text);
-    logger.debug('origin=front', { text: req.outboundMessageText }, req);
+    helpers.request.setOutboundMessageText(req, body.text);
+    req.frontConversationUrl = body._links.related.conversation;
+    logger.debug('origin=front', {
+      text: req.outboundMessageText,
+      url: req.frontConversationUrl,
+    }, req);
     const recipients = body.recipients;
+
     const fromRecipient = recipients.find(recipient => recipient.role === 'from');
     req.agentId = fromRecipient.handle;
+
     const toRecipient = recipients.find(recipient => recipient.role === 'to');
-    requestHelper.setUserId(req, toRecipient.handle);
-    req.platformMessageId = body.id;
-    req.frontConversationUrl = body._links.related.conversation;
+    try {
+      // TODO: Remove this check once all Front Conversations saved by mobile have been archived.
+      const mobileNumber = helpers.util.formatMobileNumber(toRecipient.handle);
+      helpers.request.setPlatformUserId(req, mobileNumber);
+    } catch (err) {
+      helpers.request.setUserId(req, toRecipient.handle);
+    }
   },
 };
