+++ lib/helpers/index.js
@@ -13,6 +13,7 @@ const template = require('./template');
 const twilio = require('./twilio');
 const subscription = require('./subscription');
 const user = require('./user');
+const util = require('./util');
 
 // TODO: Add "Helper" postfix to prevent name confusion with npm modules like twilio.
 module.exports = {
@@ -29,4 +30,5 @@ module.exports = {
   twilio,
   subscription,
   user,
+  util,
 };
