+++ lib/middleware/messages/message-outbound-validate.js
@@ -23,7 +23,7 @@ module.exports = function validateOutbound(config) {
       /**
        * Verify that User has valid mobile number to send a SMS.
        */
-      const mobileNumber = helpers.formatMobileNumber(req.user.mobile);
+      const mobileNumber = helpers.util.formatMobileNumber(req.user.mobile);
       helpers.request.setPlatformUserId(req, mobileNumber);
     } catch (err) {
       return helpers.sendErrorResponseWithSuppressHeaders(res, err);
